- Somes fix for new difficulty system in 5.0.5

+++ src/server/game/Entities/Creature/Creature.cpp
@@ -1059,7 +1059,7 @@ void Creature::SaveToDB()
     SaveToDB(GetMapId(), data->spawnMask, GetPhaseMask());
 }
 
-void Creature::SaveToDB(uint32 mapid, uint8 spawnMask, uint32 phaseMask)
+void Creature::SaveToDB(uint32 mapid, uint32 spawnMask, uint32 phaseMask)
 {
     // update in loaded data
     if (!m_DBTableGuid)
